# public/swagger.yaml
components:
...
    schemas:
...
        ReviewPayload:
            properties:
                shelterID:
                    type: string
                    description: 'ID of shelter'
                    example: 6475e9630044288a2b4880b5
                rating:
                    type: number
                    format: double
                    description: 'User''s given rating of shelter'
                    example: '5'
                reviewText:
                    type: string
                    description: 'User''s given review text of shelter'
                    example: 'Animal Haven Shelter provided an incredibly positive adoption experience. Highly recommended for anyone looking to adopt a pet.'
            required:
                - shelterID
                - rating
                - reviewText
            type: object
            additionalProperties: false
        ReviewResponse:
            properties:
                applicantName:
                    type: string
                applicantEmail:
                    type: string
                rating:
                    type: number
                    format: double
                reviewText:
                    type: string
            required:
                - applicantName
                - applicantEmail
                - rating
                - reviewText
            type: object
            additionalProperties: false
        ReviewsResponse:
            properties:
                reviews:
                    items:
                        $ref: '#/components/schemas/ReviewResponse'
                    type: array
                totalPages:
                    type: number
                    format: double
            required:
                - reviews
                - totalPages
            type: object
            additionalProperties: false
        ReactivationRequestPayload:
            properties:
                reasonNotScheduled:
                    type: string
                    description: 'Reason for not scheduling visit'
                    example: 'I had an unexpected personal commitment that consumed my attention during that time, and I missed the scheduling window.'
                reasonToReactivate:
                    type: string
                    description: 'Reason for requesting to reactivate'
                    example: 'The commitment has been addressed, and I am now fully available to proceed with the visit scheduling.'
                applicationID:
                    type: string
                    description: 'ID of application'
                    example: 6523d393da718fbc64f5b628
            required:
                - reasonNotScheduled
                - reasonToReactivate
                - applicationID
            type: object
            additionalProperties: false
        ReactivationRequestResponse:
            properties:
                reasonNotScheduled:
                    type: string
                    description: 'Reason for not scheduling visit'
                    example: 'I had an unexpected personal commitment that consumed my attention during that time, and I missed the scheduling window.'
                reasonToReactivate:
                    type: string
                    description: 'Reason for requesting to reactivate'
                    example: 'The commitment has been addressed, and I am now fully available to proceed with the visit scheduling.'
            required:
                - reasonNotScheduled
                - reasonToReactivate
            type: object
            additionalProperties: false
...
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
info:
    title: express-ts
    version: 1.0.0
    license:
        name: ISC
    contact: {}
openapi: 3.0.0
paths:
...
    /review:
        post:
            operationId: AddReview
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
            summary: 'Accepts rating and review and adds the review in the database'
            tags:
                - Review
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ReviewPayload'
    /review/all:
        get:
            operationId: GetReviews
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ReviewsResponse'
                            examples:
                                'Example 1':
                                    value: {reviews: [{applicantName: 'John Doe', applicantEmail: johndoe@example.com, rating: 4, reviewText: 'Great experience overall. The staff were very friendly and the place was clean and well kept.'}, {applicantName: 'Jane Smith', applicantEmail: janesmith@example.com, rating: 5, reviewText: 'I had a wonderful experience here. The staff are very professional, caring, and compassionate. I couldn''t ask for more.'}, {applicantName: 'Mark Brown', applicantEmail: markbrown@example.com, rating: 3, reviewText: 'The place was decent, but it could use a little more organization. The staff were pleasant, but they seemed understaffed.'}], totalPages: 3}
            summary: 'Returns all reviews of a specific shelter'
            tags:
                - Review
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    in: query
                    name: shelterID
                    required: true
                    schema:
                        type: string
                -
                    in: query
                    name: page
                    required: true
                    schema:
                        format: double
                        type: number
                -
                    in: query
                    name: limit
                    required: true
                    schema:
                        format: double
                        type: number
    /review/update:
        put:
            operationId: UpdateReview
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
            summary: 'Updates review'
            tags:
                - Review
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ReviewPayload'
    /reactivationRequest:
        post:
            operationId: RequestReactivation
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
            summary: 'Accepts reactivation request and sends email to shelter'
            tags:
                - ReactivationRequest
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ReactivationRequestPayload'
        get:
            operationId: GetReactivationRequest
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ReactivationRequestResponse'
                            examples:
                                'Example 1':
                                    value: {reasonNotScheduled: 'I had an unexpected personal commitment that consumed my attention during that time, and I missed the scheduling window.', reasonToReactivate: 'The commitment has been addressed, and I am now fully available to proceed with the visit scheduling.'}
            summary: 'Returns the reactivation request associated with given application ID'
            tags:
                - ReactivationRequest
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    in: query
                    name: applicationID
                    required: true
                    schema:
                        type: string
...
servers:
    -
        url: /
...